requirements.txt
google-genai
tavily-python
resend
requests
python-dotenv
arxiv
pypdf
python-slugify
